# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorPayload": {
        "additionalProperties": false,
        "description": "공개 오류 envelope 안의 공통 payload입니다.",
        "properties": {
          "code": {
            "description": "실제 HTTP response status와 같은 오류 코드입니다.",
            "title": "Code",
            "type": "integer"
          },
          "details": {
            "description": "검증 오류 상세 배열이며 일반 오류에서는 비어 있습니다.",
            "items": {
              "$ref": "#/components/schemas/ValidationDetail"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "description": "개발자용 영문 오류 설명입니다.",
            "title": "Message",
            "type": "string"
          },
          "status": {
            "description": "클라이언트가 분기할 애플리케이션 오류 식별자입니다.",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "code",
          "message",
          "details"
        ],
        "title": "ErrorPayload",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Mosemo 공개 API가 반환하는 공통 오류 response입니다.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorPayload",
            "description": "공개 오류 payload입니다."
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ValidationDetail": {
        "additionalProperties": false,
        "description": "RequestValidationError에서 선택한 공개 검증 오류 정보입니다.",
        "properties": {
          "loc": {
            "description": "Pydantic validation error의 loc 원문입니다. JSON에서는 문자열과 정수 segment의 배열입니다.",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "description": "Pydantic validation error의 msg 원문입니다.",
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "description": "Pydantic validation error의 type 원문 또는 서버 정의 오류 유형입니다.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationDetail",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Mosemo가 발급한 Bearer 액세스 토큰입니다.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Mosemo 공개 API입니다.\n\n## 공개 JSON 및 시간 계약\n\n- 공개 JSON 요청과 응답의 필드명은 camelCase를 사용합니다.\n- 성공 응답은 공통 `data`, `result`, `success` envelope 없이 도메인 데이터를 직접 반환합니다.\n- 관측 시각 요청은 `YYYY-MM-DDTHH:MM:SS[.ffffff]Z` 형식이며,\n  소수 초는 0~6자리까지 사용할 수 있습니다.\n- JSON 응답의 서버 시각은 UTC 정수 초 `YYYY-MM-DDTHH:MM:SSZ` 형식입니다.\n  Python 값에서는 마이크로초를 보존하고 JSON 직렬화에서만 반올림 없이 절삭합니다.\n- 관측 시각을 시간 블록으로 구성하는 규칙과 블록 경계는 서버가 결정합니다.\n\n## 공개 오류 계약\n\n- 애플리케이션이 생성하는 JSON 오류는 `error` 안에 `status`, `code`, `message`,\n  `details`를 담는 공통 envelope를 사용합니다.\n- `error.code`는 HTTP status와 같고, `error.status`는 클라이언트가 분기할\n  애플리케이션 오류 식별자입니다. validation detail은 RequestValidationError의\n  `loc`, `msg`, `type` 원문만 가집니다.\n- 404·405·500은 모든 v1 public operation의 공통 JSON 오류이며, 422는 실제 요청\n  입력 검증이 있는 operation에만 선언합니다.\n- 애플리케이션이 처리하지 못한 예외는 `500 INTERNAL_SERVER_ERROR` envelope로\n  반환하며, 원본 예외 정보는 응답에 포함하지 않습니다.\n- 응답 전송이 시작된 뒤 발생한 오류, proxy가 자체 생성한 오류, OAuth\n  custom-scheme redirect는 이 JSON 오류 계약의 범위 밖입니다.\n\nmacOS 생성 클라이언트 갱신은 이 서버 계약 작업의 범위 밖입니다.",
    "title": "Mosemo API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/kakao/callback": {
      "get": {
        "description": "Kakao 인증 결과와 로그인 요청의 state 및 PKCE 쿠키를 검증합니다. 성공하면 일회용 인증 코드를, 실패하면 공개 오류 코드를 macOS 앱에 전달합니다.",
        "operationId": "authKakaoCallback",
        "parameters": [
          {
            "description": "Kakao가 인증 성공 시 전달한 authorization code입니다.",
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Kakao가 인증 성공 시 전달한 authorization code입니다.",
              "title": "Code"
            }
          },
          {
            "description": "로그인 요청과 callback을 연결하고 CSRF를 방지하는 state 값입니다.",
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "로그인 요청과 callback을 연결하고 CSRF를 방지하는 state 값입니다.",
              "title": "State"
            }
          },
          {
            "description": "Kakao가 인증 실패 또는 취소 시 전달한 오류 코드입니다.",
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Kakao가 인증 실패 또는 취소 시 전달한 오류 코드입니다.",
              "title": "Error"
            }
          },
          {
            "description": "Kakao가 전달한 상세 오류 설명입니다. 서버는 민감 정보 노출을 방지하기 위해 사용하지 않습니다.",
            "in": "query",
            "name": "error_description",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Kakao가 전달한 상세 오류 설명입니다. 서버는 민감 정보 노출을 방지하기 위해 사용하지 않습니다.",
              "title": "Error Description"
            }
          },
          {
            "description": "로그인 시작 시 서버가 저장한 OAuth state 쿠키입니다.",
            "in": "cookie",
            "name": "kakao_oauth_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "로그인 시작 시 서버가 저장한 OAuth state 쿠키입니다.",
              "title": "Kakao Oauth State"
            }
          },
          {
            "description": "로그인 시작 시 서버가 저장한 PKCE code challenge 쿠키입니다.",
            "in": "cookie",
            "name": "kakao_oauth_pkce_challenge",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "로그인 시작 시 서버가 저장한 PKCE code challenge 쿠키입니다.",
              "title": "Kakao Oauth Pkce Challenge"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "인증 결과를 macOS 앱에 전달하는 리다이렉트 응답입니다.",
            "headers": {
              "Cache-Control": {
                "description": "응답을 저장하지 않도록 no-store로 설정됩니다.",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "Location": {
                "description": "성공 시 code, 실패 시 access_denied 또는 authentication_failed를 포함하는 macOS custom scheme URI입니다.",
                "schema": {
                  "format": "uri",
                  "type": "string"
                }
              },
              "Pragma": {
                "description": "이전 HTTP 캐시와의 호환을 위해 no-cache로 설정됩니다.",
                "schema": {
                  "const": "no-cache",
                  "type": "string"
                }
              },
              "Set-Cookie": {
                "description": "state 및 PKCE challenge 쿠키를 즉시 만료시킵니다.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "AUTH_INVALID_OAUTH_CONTEXT": {
                    "description": "OAuth login context가 유효하지 않습니다.",
                    "summary": "Invalid OAuth context",
                    "value": {
                      "error": {
                        "code": 400,
                        "details": [],
                        "message": "Invalid or expired OAuth login context",
                        "status": "AUTH_INVALID_OAUTH_CONTEXT"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request",
            "headers": {
              "Cache-Control": {
                "description": "응답을 저장하지 않도록 no-store로 설정됩니다.",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "Pragma": {
                "description": "이전 HTTP 캐시와의 호환을 위해 no-cache로 설정됩니다.",
                "schema": {
                  "const": "no-cache",
                  "type": "string"
                }
              },
              "Set-Cookie": {
                "description": "state 및 PKCE challenge 쿠키를 즉시 만료시킵니다.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "REQUEST_ROUTE_NOT_FOUND": {
                    "description": "요청한 API route를 찾을 수 없습니다.",
                    "summary": "Route not found",
                    "value": {
                      "error": {
                        "code": 404,
                        "details": [],
                        "message": "API route not found",
                        "status": "REQUEST_ROUTE_NOT_FOUND"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found",
            "headers": {}
          },
          "405": {
            "content": {
              "application/json": {
                "examples": {
                  "REQUEST_METHOD_NOT_ALLOWED": {
                    "description": "요청 method가 해당 API route에서 허용되지 않습니다.",
                    "summary": "Method not allowed",
                    "value": {
                      "error": {
                        "code": 405,
                        "details": [],
                        "message": "Method not allowed",
                        "status": "REQUEST_METHOD_NOT_ALLOWED"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method Not Allowed",
            "headers": {
              "Allow": {
                "description": "해당 route에서 허용되는 HTTP method 목록입니다.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "INTERNAL_SERVER_ERROR": {
                    "description": "예상하지 못한 서버 오류가 발생했습니다.",
                    "summary": "Internal server error",
                    "value": {
                      "error": {
                        "code": 500,
                        "details": [],
                        "message": "Internal server error",
                        "status": "INTERNAL_SERVER_ERROR"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error",
            "headers": {}
          }
        },
        "summary": "Kakao 로그인 callback 처리",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/kakao/login": {
      "get": {
        "description": "macOS 앱이 생성한 PKCE code challenge를 저장하고 Kakao 인증 화면으로 이동시킵니다.",
        "operationId": "authKakaoLogin",
        "parameters": [
          {
            "description": "macOS 앱이 code_verifier로부터 S256 방식으로 생성한 PKCE code challenge입니다.",
            "in": "query",
            "name": "code_challenge",
            "required": true,
            "schema": {
              "description": "macOS 앱이 code_verifier로부터 S256 방식으로 생성한 PKCE code challenge입니다.",
              "pattern": "^[A-Za-z0-9_-]{43}$",
              "title": "Code Challenge",
              "type": "string"
            }
          },
          {
            "description": "PKCE code challenge 생성 방식이며 S256으로 고정됩니다.",
            "in": "query",
            "name": "code_challenge_method",
            "required": true,
            "schema": {
              "const": "S256",
              "description": "PKCE code challenge 생성 방식이며 S256으로 고정됩니다.",
              "title": "Code Challenge Method",
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Kakao 인증 화면으로 이동하는 리다이렉트 응답입니다.",
            "headers": {
              "Cache-Control": {
                "description": "응답을 저장하지 않도록 no-store로 설정됩니다.",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "Location": {
                "description": "Kakao 인증 화면의 URL입니다.",
                "schema": {
                  "format": "uri",
                  "type": "string"
                }
              },
              "Pragma": {
                "description": "이전 HTTP 캐시와의 호환을 위해 no-cache로 설정됩니다.",
                "schema": {
                  "const": "no-cache",
                  "type": "string"
                }
              },
              "Set-Cookie": {
                "description": "10분 동안 유지되는 HttpOnly state 및 PKCE challenge 쿠키입니다. 운영 환경에서는 Secure 속성도 사용합니다.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "REQUEST_ROUTE_NOT_FOUND": {
                    "description": "요청한 API route를 찾을 수 없습니다.",
                    "summary": "Route not found",
                    "value": {
                      "error": {
                        "code": 404,
                        "details": [],
                        "message": "API route not found",
                        "status": "REQUEST_ROUTE_NOT_FOUND"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found",
            "headers": {}
          },
          "405": {
            "content": {
              "application/json": {
                "examples": {
                  "REQUEST_METHOD_NOT_ALLOWED": {
                    "description": "요청 method가 해당 API route에서 허용되지 않습니다.",
                    "summary": "Method not allowed",
                    "value": {
                      "error": {
                        "code": 405,
                        "details": [],
                        "message": "Method not allowed",
                        "status": "REQUEST_METHOD_NOT_ALLOWED"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method Not Allowed",
            "headers": {
              "Allow": {
                "description": "해당 route에서 허용되는 HTTP method 목록입니다.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "INVALID_ARGUMENT": {
                    "description": "요청 입력값 검증에 실패했습니다.",
                    "summary": "Invalid argument",
                    "value": {
                      "error": {
                        "code": 422,
                        "details": [
                          {
                            "loc": [
                              "body",
                              "users",
                              0,
                              "email"
                            ],
                            "msg": "Field required",
                            "type": "missing"
                          }
                        ],
                        "message": "Request validation failed.",
                        "status": "INVALID_ARGUMENT"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content",
            "headers": {}
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "INTERNAL_SERVER_ERROR": {
                    "description": "예상하지 못한 서버 오류가 발생했습니다.",
                    "summary": "Internal server error",
                    "value": {
                      "error": {
                        "code": 500,
                        "details": [],
                        "message": "Internal server error",
                        "status": "INTERNAL_SERVER_ERROR"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error",
            "headers": {}
          }
        },
        "summary": "Kakao 로그인 시작",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}